+++ src/jetbrains/buildServer/server/rest/request/ProjectRequest.java
@@ -109,7 +109,7 @@ public class ProjectRequest {
                             @PathParam("btLocator") String buildTypeLocator,
                             @QueryParam("status") String status,
                             @QueryParam("start") @DefaultValue(value = "0") Long start,
-                            @QueryParam("count") @DefaultValue(value = "100") Long count) {
+                            @QueryParam("count") @DefaultValue(value = "100") Integer count) {
     SBuildType buildType = myDataProvider.getBuildType(myDataProvider.getProject(projectLocator), buildTypeLocator);
 
     final List<SFinishedBuild> buildsList = myDataProvider.getBuilds(buildType, null, false, true, false, status, start, count);
